Removed bug in nextbatch.m
It is possible to reach the bottom of pool points and not have filled the batch matrix to the requested nbatch points.
this is due to some batch points being disquilified, and reaching the bottom of pool quicker than batch is filled.
this would throw an error which is now fixed: Batch is re-adjusted.

Changed some print to screen texts
Renamed theta07 file in optimum theta folder