Changed batch pool creation method and prediction points for verify
Pool points are now the same pred points which fall inside the user specified boundary. Before, pool points were created from scratch using
Halton sequences but now they are simply the same points as pred
Prediction points created for verify used to be linspace in XY but are now Halton. One set of pred points is created using Halton and then
if we're on the verify objective, all SA values are replaced with nominal. If we're on batch then nothing is replaced
Prediction MSE's smaller than eps are set to 0
Since pool is changed then some interpx and interpy inside plotgek.m is also changed accordingly